refactor: repository 패키지 분리 및 메서드 정렬 규칙 지정
