+++ src/main/java/roomescape/business/vo/Period.java
@@ -1,0 +1,9 @@
+package roomescape.business.vo;
+
+import java.time.LocalDate;
+
+public record Period(
+    LocalDate startDate,
+    LocalDate endDate) {
+
+}
