+++ src/main/java/roomescape/infrastructure/db/TossPaymentJpaRepository.java
@@ -4,4 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import roomescape.business.model.TossPayment;
 
 public interface TossPaymentJpaRepository extends JpaRepository<TossPayment, Long> {
+
 }
