+++ src/main/java/roomescape/infrastructure/db/WaitingJpaRepository.java
@@ -13,31 +13,29 @@ import roomescape.business.model.Waiting;
 
 public interface WaitingJpaRepository extends JpaRepository<Waiting, Long> {
 
-    List<Waiting> findByReservation_MemberId(Long id);
+    List<Waiting> findByReservation_MemberId(Long memberId);
 
     @Query("""
-             SELECT COUNT(w) FROM Waiting w
-             WHERE w.registeredAt < :registeredAt AND 
-                     w.reservation.date = :date AND
-                    w.reservation.theme = :theme AND 
-                    w.reservation.reservationTime = :reservationTime
-            """)
+        SELECT w FROM Waiting w
+                WHERE  w.reservation.date = :date AND
+                w.reservation.reservationTime = :reservationTime AND
+                w.reservation.theme = :theme
+        ORDER BY w.registeredAt ASC
+        LIMIT 1
+        """)
+    Optional<Waiting> findEarliestWaitingBy(
+        LocalDate date, ReservationTime reservationTime, Theme theme);
+
+    @Query("""
+         SELECT COUNT(w) FROM Waiting w
+         WHERE w.registeredAt < :registeredAt AND
+                 w.reservation.date = :date AND
+                w.reservation.theme = :theme AND
+                w.reservation.reservationTime = :reservationTime
+        """)
     int countWaitingBefore(
-            @Param("registeredAt") LocalDateTime registeredAt,
-            @Param("date") LocalDate date,
-            @Param("theme") Theme theme,
-            @Param("reservationTime") ReservationTime reservationTime
-    );
-
-
-    @Query("""
-            SELECT w FROM Waiting w
-                    WHERE  w.reservation.date = :date AND
-                    w.reservation.reservationTime = :reservationTime AND
-                    w.reservation.theme = :theme
-            ORDER BY w.registeredAt ASC
-            LIMIT 1
-            """)
-    Optional<Waiting> findEarliestWaitingBy(LocalDate date, ReservationTime reservationTime, Theme theme);
+        @Param("registeredAt") LocalDateTime registeredAt,
+        @Param("date") LocalDate date,
+        @Param("theme") Theme theme,
+        @Param("reservationTime") ReservationTime reservationTime);
 }
-
