+++ src/main/java/roomescape/persistence/MemberRepository.java
@@ -1,4 +1,4 @@
-package roomescape.persistence.repository;
+package roomescape.persistence;
 
 import roomescape.business.model.Member;
 
@@ -7,5 +7,4 @@ public interface MemberRepository {
     Member findByEmail(String email);
 
     Member findById(Long id);
-
 }
