+++ src/main/java/roomescape/persistence/TossPaymentRepository.java
@@ -1,7 +1,8 @@
-package roomescape.persistence.repository;
+package roomescape.persistence;
 
 import roomescape.business.model.TossPayment;
 
 public interface TossPaymentRepository {
+
     TossPayment save(TossPayment tossPayment);
 }
