+++ src/main/java/roomescape/persistence/implementation/MemberRepositoryImpl.java
@@ -1,10 +1,11 @@
-package roomescape.persistence.repository;
+package roomescape.persistence.implementation;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
+import roomescape.business.model.Member;
 import roomescape.common.exception.NotFoundException;
 import roomescape.infrastructure.db.MemberJpaRepository;
-import roomescape.business.model.Member;
+import roomescape.persistence.MemberRepository;
 
 @Repository
 @RequiredArgsConstructor
@@ -15,12 +16,12 @@ public class MemberRepositoryImpl implements MemberRepository {
     @Override
     public Member findByEmail(String email) {
         return memberJpaRepository.findByEmail(email)
-                .orElseThrow(() -> new NotFoundException("존재하지 않는 사용자입니다."));
+            .orElseThrow(() -> new NotFoundException("존재하지 않는 사용자입니다."));
     }
 
     @Override
     public Member findById(Long id) {
         return memberJpaRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("존재하지 않는 사용자입니다."));
+            .orElseThrow(() -> new NotFoundException("존재하지 않는 사용자입니다."));
     }
 }
