+++ src/main/java/roomescape/persistence/implementation/ReservationTimeRepositoryImpl.java
@@ -1,12 +1,13 @@
-package roomescape.persistence.repository;
+package roomescape.persistence.implementation;
 
 import java.time.LocalTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
+import roomescape.business.model.ReservationTime;
 import roomescape.common.exception.NotFoundException;
 import roomescape.infrastructure.db.ReservationTimeJpaRepository;
-import roomescape.business.model.ReservationTime;
+import roomescape.persistence.ReservationTimeRepository;
 
 @Repository
 @RequiredArgsConstructor
@@ -15,19 +16,19 @@ public class ReservationTimeRepositoryImpl implements ReservationTimeRepository 
     private final ReservationTimeJpaRepository reservationTimeJpaRepository;
 
     @Override
-    public boolean isDuplicatedStartAt(LocalTime startAt) {
-        return reservationTimeJpaRepository.existsByStartAt(startAt);
+    public List<ReservationTime> findAll() {
+        return reservationTimeJpaRepository.findAll();
     }
 
     @Override
     public ReservationTime findById(Long id) {
         return reservationTimeJpaRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("존재하지 않는 예약 시각입니다."));
+            .orElseThrow(() -> new NotFoundException("존재하지 않는 예약 시각입니다."));
     }
 
     @Override
-    public List<ReservationTime> findAll() {
-        return reservationTimeJpaRepository.findAll();
+    public boolean isDuplicatedStartAt(LocalTime startAt) {
+        return reservationTimeJpaRepository.existsByStartAt(startAt);
     }
 
     @Override
