+++ src/main/java/roomescape/persistence/implementation/TossPaymentRepositoryImpl.java
@@ -1,18 +1,19 @@
-package roomescape.persistence.repository;
+package roomescape.persistence.implementation;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import roomescape.infrastructure.db.TossPaymentJpaRepository;
 import roomescape.business.model.TossPayment;
+import roomescape.infrastructure.db.TossPaymentJpaRepository;
+import roomescape.persistence.TossPaymentRepository;
 
 @Repository
 @RequiredArgsConstructor
 public class TossPaymentRepositoryImpl implements TossPaymentRepository {
 
-    private final TossPaymentJpaRepository jpaRepository;
+    private final TossPaymentJpaRepository tossPaymentJpaRepository;
 
     @Override
     public TossPayment save(TossPayment tossPayment) {
-        return jpaRepository.save(tossPayment);
+        return tossPaymentJpaRepository.save(tossPayment);
     }
 }
