--- src/main/java/roomescape/persistence/repository/ThemeRepository.java
@@ -1,23 +1,0 @@
-package roomescape.persistence.repository;
-
-import java.util.List;
-import roomescape.business.model.Theme;
-import roomescape.persistence.vo.Period;
-
-public interface ThemeRepository {
-
-    boolean isDuplicatedName(String name);
-
-    List<Theme> findPopularThemesInPeriod(
-            Period period,
-            int size
-    );
-
-    Theme findById(Long id);
-
-    List<Theme> findAll();
-
-    Theme save(Theme theme);
-
-    void deleteById(Long id);
-}
