--- src/main/java/roomescape/persistence/vo/Period.java
@@ -1,9 +1,0 @@
-package roomescape.persistence.vo;
-
-import java.time.LocalDate;
-
-public record Period(
-        LocalDate startDate,
-        LocalDate endDate
-) {
-}
