+++ src/test/java/roomescape/presentation/resvoler/LoginMemberArgumentResolverTest.java
@@ -16,7 +16,7 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.context.request.ServletWebRequest;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import roomescape.dto.LoginMember;
+import roomescape.business.vo.LoginMember;
 import roomescape.business.model.Member;
 import roomescape.business.model.Role;
 import roomescape.presentation.support.CookieUtils;
